reup

+++ src/Image.cpp
@@ -8,9 +8,9 @@ static const int ascii_lum_length = 73;
  Image::Image(unsigned int WIDTH, unsigned int HEIGHT, unsigned int LABEL) : 
     WIDTH(WIDTH), HEIGHT(HEIGHT), LABEL(LABEL) {
 
-    for(int y = 0; y < HEIGHT; y++){
+    for(unsigned int y = 0; y < HEIGHT; y++){
         pixel_values.push_back(std::vector<unsigned int>());
-        for(int x = 0; x < WIDTH; x++){
+        for(unsigned int x = 0; x < WIDTH; x++){
             pixel_values[y].push_back(0);
         }
     }
@@ -22,19 +22,19 @@ void Image::put(int x, int y, int val){
 }
 
 float Image::get(int x, int y){
-    return pixel_values[y][x];
+    return (float)pixel_values[y][x];
 }
 
 void Image::print(){
     std::cout << std::endl << "IMAGE OF A <" << LABEL << ">:" << std::endl;
 
     //border
-    for(int i = 0; i <= (WIDTH / 2); i++) printf("+ ");
+    for(unsigned int i = 0; i <= (WIDTH / 2); i++) printf("+ ");
     printf("\n");
 
-    for(int y = 0; y < HEIGHT; y++){
+    for(unsigned int y = 0; y < HEIGHT; y++){
         std::string line = "+"; //border
-        for(int x = 0; x < WIDTH; x++){
+        for(unsigned int x = 0; x < WIDTH; x++){
             int val = pixel_values[y][x];
             float brightness = fmax(fmin((float(val)/255.0f),1.0f),0.0f);
